+++ custom_components/lutarym_ppc_smgw/__init__.py
@@ -1,4 +1,4 @@
-# Integrationsversion: 2.4.5
+# Integrationsversion: 2.4.6
 """PPC Smart Meter Gateway (iMSys) Integration für Home Assistant.
 
 Einstiegspunkt der Integration (von Home Assistant automatisch anhand des